feat(Core): ESI API rate limits support
fixes #534
