+++ lib/wanderer_app/esi/api_client.ex
@@ -6,35 +6,9 @@ defmodule WandererApp.Esi.ApiClient do
   alias WandererApp.Cache
 
   @ttl :timer.hours(1)
-  @routes_ttl :timer.minutes(15)
 
-  @base_url "https://esi.evetech.net/latest"
   @wanderrer_user_agent "(wanderer-industries@proton.me; +https://github.com/wanderer-industries/wanderer)"
-
-  @req_esi Req.new(base_url: @base_url, finch: WandererApp.Finch)
-
-  @get_link_pairs_advanced_params [
-    :include_mass_crit,
-    :include_eol,
-    :include_frig
-  ]
-
-  @default_routes_settings %{
-    path_type: "shortest",
-    include_mass_crit: true,
-    include_eol: false,
-    include_frig: true,
-    include_cruise: true,
-    avoid_wormholes: false,
-    avoid_pochven: false,
-    avoid_edencom: false,
-    avoid_triglavian: false,
-    include_thera: true,
-    avoid: []
-  }
-
-  @zarzakh_system 30_100_000
-  @default_avoid_systems [@zarzakh_system]
+  @req_esi_options [base_url: "https://esi.evetech.net", finch: WandererApp.Finch]
 
   @cache_opts [cache: true]
   @retry_opts [retry: false, retry_log_level: :warning]
@@ -43,11 +17,11 @@ defmodule WandererApp.Esi.ApiClient do
 
   @logger Application.compile_env(:wanderer_app, :logger)
 
-  def get_server_status, do: get("/status")
+  def get_server_status, do: do_get("/status", [], @cache_opts)
 
   def set_autopilot_waypoint(add_to_beginning, clear_other_waypoints, destination_id, opts \\ []),
     do:
-      post_esi(
+      do_post_esi(
         "/ui/autopilot/waypoint",
         get_auth_opts(opts)
         |> Keyword.merge(
@@ -62,7 +36,7 @@ defmodule WandererApp.Esi.ApiClient do
   def post_characters_affiliation(character_eve_ids, _opts)
       when is_list(character_eve_ids),
       do:
-        post_esi(
+        do_post_esi(
           "/characters/affiliation/",
           json: character_eve_ids,
           params: %{
@@ -70,168 +44,9 @@ defmodule WandererApp.Esi.ApiClient do
           }
         )
 
-  def find_routes(map_id, origin, hubs, routes_settings) do
-    origin = origin |> String.to_integer()
-    hubs = hubs |> Enum.map(&(&1 |> String.to_integer()))
-
-    routes_settings = @default_routes_settings |> Map.merge(routes_settings)
-
-    connections =
-      case routes_settings.avoid_wormholes do
-        false ->
-          map_chains =
-            routes_settings
-            |> Map.take(@get_link_pairs_advanced_params)
-            |> Map.put_new(:map_id, map_id)
-            |> WandererApp.Api.MapConnection.get_link_pairs_advanced!()
-            |> Enum.map(fn %{
-                             solar_system_source: solar_system_source,
-                             solar_system_target: solar_system_target
-                           } ->
-              %{
-                first: solar_system_source,
-                second: solar_system_target
-              }
-            end)
-            |> Enum.uniq()
-
-          {:ok, thera_chains} =
-            case routes_settings.include_thera do
-              true ->
-                WandererApp.Server.TheraDataFetcher.get_chain_pairs(routes_settings)
-
-              false ->
-                {:ok, []}
-            end
-
-          chains = remove_intersection([map_chains | thera_chains] |> List.flatten())
-
-          chains =
-            case routes_settings.include_cruise do
-              false ->
-                {:ok, wh_class_a_systems} = WandererApp.CachedInfo.get_wh_class_a_systems()
-
-                chains
-                |> Enum.filter(fn x ->
-                  not Enum.member?(wh_class_a_systems, x.first) and
-                    not Enum.member?(wh_class_a_systems, x.second)
-                end)
-
-              _ ->
-                chains
-            end
-
-          chains
-          |> Enum.map(fn chain ->
-            ["#{chain.first}|#{chain.second}", "#{chain.second}|#{chain.first}"]
-          end)
-          |> List.flatten()
-
-        true ->
-          []
-      end
-
-    {:ok, trig_systems} = WandererApp.CachedInfo.get_trig_systems()
-
-    pochven_solar_systems =
-      trig_systems
-      |> Enum.filter(fn s -> s.triglavian_invasion_status == "Final" end)
-      |> Enum.map(& &1.solar_system_id)
-
-    triglavian_solar_systems =
-      trig_systems
-      |> Enum.filter(fn s -> s.triglavian_invasion_status == "Triglavian" end)
-      |> Enum.map(& &1.solar_system_id)
-
-    edencom_solar_systems =
-      trig_systems
-      |> Enum.filter(fn s -> s.triglavian_invasion_status == "Edencom" end)
-      |> Enum.map(& &1.solar_system_id)
-
-    avoidance_list =
-      case routes_settings.avoid_edencom do
-        true ->
-          edencom_solar_systems
-
-        false ->
-          []
-      end
-
-    avoidance_list =
-      case routes_settings.avoid_triglavian do
-        true ->
-          [avoidance_list | triglavian_solar_systems]
-
-        false ->
-          avoidance_list
-      end
-
-    avoidance_list =
-      case routes_settings.avoid_pochven do
-        true ->
-          [avoidance_list | pochven_solar_systems]
-
-        false ->
-          avoidance_list
-      end
-
-    avoidance_list =
-      (@default_avoid_systems ++ [routes_settings.avoid | avoidance_list])
-      |> List.flatten()
-      |> Enum.uniq()
-
-    params =
-      %{
-        datasource: "tranquility",
-        flag: routes_settings.path_type,
-        connections: connections,
-        avoid: avoidance_list
-      }
-
-    {:ok, all_routes} = get_all_routes(hubs, origin, params)
-
-    routes =
-      all_routes
-      |> Enum.map(fn route_info ->
-        map_route_info(route_info)
-      end)
-      |> Enum.filter(fn route_info -> not is_nil(route_info) end)
-
-    {:ok, routes}
-  end
-
-  def get_all_routes(hubs, origin, params, opts \\ []) do
-    cache_key =
-      "routes-#{origin}-#{hubs |> Enum.join("-")}-#{:crypto.hash(:sha, :erlang.term_to_binary(params))}"
-
-    case WandererApp.Cache.lookup(cache_key) do
-      {:ok, result} when not is_nil(result) ->
-        {:ok, result}
-
-      _ ->
-        case get_all_routes_custom(hubs, origin, params) do
-          {:ok, result} ->
-            WandererApp.Cache.insert(
-              cache_key,
-              result,
-              ttl: @routes_ttl
-            )
-
-            {:ok, result}
-
-          {:error, _error} ->
-            @logger.error(
-              "Error getting custom routes for #{inspect(origin)}: #{inspect(params)}"
-            )
-
-            get_all_routes_eve(hubs, origin, params, opts)
-        end
-    end
-  end
-
-  defp get_all_routes_custom(hubs, origin, params),
+  def get_routes_custom(hubs, origin, params),
     do:
-      post(
+      do_post(
         "#{get_custom_route_base_url()}/route/multiple",
         [
           json: %{
@@ -245,13 +60,20 @@ defmodule WandererApp.Esi.ApiClient do
         |> Keyword.merge(@timeout_opts)
       )
 
-  def get_all_routes_eve(hubs, origin, params, opts),
+  def get_routes_eve(hubs, origin, params, opts),
     do:
       {:ok,
        hubs
        |> Task.async_stream(
          fn destination ->
-           get_routes(origin, destination, params, opts)
+           %{
+             "origin" => origin,
+             "destination" => destination,
+             "systems" => [],
+             "success" => false
+           }
+
+           # do_get_routes_eve(origin, destination, params, opts)
          end,
          max_concurrency: System.schedulers_online() * 4,
          timeout: :timer.seconds(30),
@@ -265,8 +87,19 @@ defmodule WandererApp.Esi.ApiClient do
          end
        end)}
 
-  def get_routes(origin, destination, params, opts) do
-    case _get_routes(origin, destination, params, opts) do
+  defp do_get_routes_eve(origin, destination, params, opts) do
+    esi_params =
+      Map.merge(params, %{
+        connections: params.connections |> Enum.join(","),
+        avoid: params.avoid |> Enum.join(",")
+      })
+
+    do_get(
+      "/route/#{origin}/#{destination}/?#{esi_params |> Plug.Conn.Query.encode()}",
+      opts,
+      @cache_opts
+    )
+    |> case do
       {:ok, result} ->
         %{
           "origin" => origin,
@@ -299,9 +132,8 @@ defmodule WandererApp.Esi.ApiClient do
               key: "killmail-#{killmail_id}-#{killmail_hash}",
               opts: [ttl: @ttl]
             )
-  def get_killmail(killmail_id, killmail_hash, opts \\ []) do
-    get("/killmails/#{killmail_id}/#{killmail_hash}/", opts, @cache_opts)
-  end
+  def get_killmail(killmail_id, killmail_hash, opts \\ []),
+    do: do_get("/killmails/#{killmail_id}/#{killmail_hash}/", opts, @cache_opts)
 
   @decorate cacheable(
               cache: Cache,
@@ -322,7 +154,7 @@ defmodule WandererApp.Esi.ApiClient do
               opts: [ttl: @ttl]
             )
   def get_character_info(eve_id, opts \\ []) do
-    case get(
+    case do_get(
            "/characters/#{eve_id}/",
            opts,
            @cache_opts
@@ -395,48 +227,11 @@ defmodule WandererApp.Esi.ApiClient do
     get_character_auth_data(character_eve_id, "search", merged_opts)
   end
 
-  defp remove_intersection(pairs_arr) do
-    tuples = pairs_arr |> Enum.map(fn x -> {x.first, x.second} end)
-
-    tuples
-    |> Enum.reduce([], fn {first, second} = x, acc ->
-      if Enum.member?(tuples, {second, first}) do
-        acc
-      else
-        [x | acc]
-      end
-    end)
-    |> Enum.uniq()
-    |> Enum.map(fn {first, second} ->
-      %{
-        first: first,
-        second: second
-      }
-    end)
-  end
-
-  defp _get_routes(origin, destination, params, opts),
-    do: get_routes_eve(origin, destination, params, opts)
-
-  defp get_routes_eve(origin, destination, params, opts) do
-    esi_params =
-      Map.merge(params, %{
-        connections: params.connections |> Enum.join(","),
-        avoid: params.avoid |> Enum.join(",")
-      })
-
-    get(
-      "/route/#{origin}/#{destination}/?#{esi_params |> Plug.Conn.Query.encode()}",
-      opts,
-      @cache_opts
-    )
-  end
-
   defp get_auth_opts(opts), do: [auth: {:bearer, opts[:access_token]}]
 
   defp get_alliance_info(alliance_eve_id, info_path, opts),
     do:
-      get(
+      do_get(
         "/alliances/#{alliance_eve_id}/#{info_path}",
         opts,
         @cache_opts
@@ -444,7 +239,7 @@ defmodule WandererApp.Esi.ApiClient do
 
   defp get_corporation_info(corporation_eve_id, info_path, opts),
     do:
-      get(
+      do_get(
         "/corporations/#{corporation_eve_id}/#{info_path}",
         opts,
         @cache_opts
@@ -460,13 +255,13 @@ defmodule WandererApp.Esi.ApiClient do
     character_id = opts |> Keyword.get(:character_id, nil)
 
     if not is_access_token_expired?(character_id) do
-      get(
+      do_get(
         path,
         auth_opts,
         opts |> with_refresh_token()
       )
     else
-      get_retry(path, auth_opts, opts |> with_refresh_token())
+      do_get_retry(path, auth_opts, opts |> with_refresh_token())
     end
   end
 
@@ -481,29 +276,26 @@ defmodule WandererApp.Esi.ApiClient do
 
   defp get_corporation_auth_data(corporation_eve_id, info_path, opts),
     do:
-      get(
+      do_get(
         "/corporations/#{corporation_eve_id}/#{info_path}",
         [params: opts[:params] || []] ++
           (opts |> get_auth_opts()),
         (opts |> with_refresh_token()) ++ @cache_opts
       )
 
-  defp with_user_agent_opts(opts) do
-    opts
-    |> Keyword.merge(
-      headers: [{:user_agent, "Wanderer/#{WandererApp.Env.vsn()} #{@wanderrer_user_agent}"}]
-    )
-  end
+  defp with_user_agent_opts(opts),
+    do:
+      opts
+      |> Keyword.merge(
+        headers: [{:user_agent, "Wanderer/#{WandererApp.Env.vsn()} #{@wanderrer_user_agent}"}]
+      )
 
-  defp with_refresh_token(opts) do
-    opts |> Keyword.merge(refresh_token?: true)
-  end
+  defp with_refresh_token(opts), do: opts |> Keyword.merge(refresh_token?: true)
 
-  defp with_cache_opts(opts) do
-    opts |> Keyword.merge(@cache_opts) |> Keyword.merge(cache_dir: System.tmp_dir!())
-  end
+  defp with_cache_opts(opts),
+    do: opts |> Keyword.merge(@cache_opts) |> Keyword.merge(cache_dir: System.tmp_dir!())
 
-  defp get(path, api_opts \\ [], opts \\ []) do
+  defp do_get(path, api_opts \\ [], opts \\ []) do
     case Cachex.get(:api_cache, path) do
       {:ok, cached_data} when not is_nil(cached_data) ->
         {:ok, cached_data}
@@ -515,15 +307,17 @@ defmodule WandererApp.Esi.ApiClient do
 
   defp do_get_request(path, api_opts \\ [], opts \\ []) do
     try do
-      case Req.get(
-             @req_esi,
-             api_opts
-             |> Keyword.merge(url: path)
-             |> with_user_agent_opts()
-             |> with_cache_opts()
-             |> Keyword.merge(@retry_opts)
-             |> Keyword.merge(@timeout_opts)
-           ) do
+      @req_esi_options
+      |> Req.new()
+      |> Req.get(
+        api_opts
+        |> Keyword.merge(url: path)
+        |> with_user_agent_opts()
+        |> with_cache_opts()
+        |> Keyword.merge(@retry_opts)
+        |> Keyword.merge(@timeout_opts)
+      )
+      |> case do
         {:ok, %{status: 200, body: body, headers: headers}} ->
           maybe_cache_response(path, body, headers, opts)
 
@@ -537,8 +331,8 @@ defmodule WandererApp.Esi.ApiClient do
 
         {:ok, %{status: 420, headers: headers} = _error} ->
           # Extract rate limit information from headers
-          reset_seconds = Map.get(headers, "x-esi-error-limit-reset", ["unknown"]) |> List.first()
-          remaining = Map.get(headers, "x-esi-error-limit-remain", ["unknown"]) |> List.first()
+          reset_seconds = Map.get(headers, "x-esi-error-limit-reset", ["0"]) |> List.first()
+          remaining = Map.get(headers, "x-esi-error-limit-remain", ["0"]) |> List.first()
 
           # Emit telemetry for rate limiting
           :telemetry.execute(
@@ -568,10 +362,40 @@ defmodule WandererApp.Esi.ApiClient do
 
           {:error, :error_limited, headers}
 
-        {:ok, %{status: status} = _error} when status in [401, 403] ->
-          get_retry(path, api_opts, opts)
+        {:ok, %{status: 429, headers: headers} = _error} ->
+          # Extract rate limit information from headers
+          reset_seconds = Map.get(headers, "retry-after", ["0"]) |> List.first()
 
-        {:ok, %{status: status}} ->
+          # Emit telemetry for rate limiting
+          :telemetry.execute(
+            [:wanderer_app, :esi, :rate_limited],
+            %{
+              count: 1,
+              reset_duration:
+                case Integer.parse(reset_seconds || "0") do
+                  {seconds, _} -> seconds * 1000
+                  _ -> 0
+                end
+            },
+            %{
+              method: "GET",
+              path: path,
+              reset_seconds: reset_seconds
+            }
+          )
+
+          Logger.warning("ESI_RATE_LIMITED: GET request rate limited",
+            method: "GET",
+            path: path,
+            reset_seconds: reset_seconds
+          )
+
+          {:error, :error_limited, headers}
+
+        {:ok, %{status: status} = _error} when status in [401, 403] ->
+          do_get_retry(path, api_opts, opts)
+
+        {:ok, %{status: status, headers: headers}} ->
           {:error, "Unexpected status: #{status}"}
 
         {:error, _reason} ->
@@ -585,7 +409,7 @@ defmodule WandererApp.Esi.ApiClient do
     end
   end
 
-  defp maybe_cache_response(path, body, %{"expires" => [expires]}, opts)
+  defp maybe_cache_response(path, body, %{"expires" => [expires]} = _headers, opts)
        when is_binary(path) and not is_nil(expires) do
     try do
       if opts |> Keyword.get(:cache, false) do
@@ -609,7 +433,7 @@ defmodule WandererApp.Esi.ApiClient do
 
   defp maybe_cache_response(_path, _body, _headers, _opts), do: :ok
 
-  defp post(url, opts) do
+  defp do_post(url, opts) do
     try do
       case Req.post("#{url}", opts |> with_user_agent_opts()) do
         {:ok, %{status: status, body: body}} when status in [200, 201] ->
@@ -623,8 +447,8 @@ defmodule WandererApp.Esi.ApiClient do
 
         {:ok, %{status: 420, headers: headers} = _error} ->
           # Extract rate limit information from headers
-          reset_seconds = Map.get(headers, "x-esi-error-limit-reset", ["unknown"]) |> List.first()
-          remaining = Map.get(headers, "x-esi-error-limit-remain", ["unknown"]) |> List.first()
+          reset_seconds = Map.get(headers, "x-esi-error-limit-reset", ["0"]) |> List.first()
+          remaining = Map.get(headers, "x-esi-error-limit-remain", ["0"]) |> List.first()
 
           # Emit telemetry for rate limiting
           :telemetry.execute(
@@ -668,16 +492,13 @@ defmodule WandererApp.Esi.ApiClient do
     end
   end
 
-  defp post_esi(url, opts) do
+  defp do_post_esi(url, opts) do
     try do
       req_opts =
         (opts |> with_user_agent_opts() |> Keyword.merge(@retry_opts)) ++
           [params: opts[:params] || []]
 
-      Req.new(
-        [base_url: @base_url, finch: WandererApp.Finch] ++
-          req_opts
-      )
+      Req.new(@req_esi_options ++ req_opts)
       |> Req.post(url: url)
       |> case do
         {:ok, %{status: status, body: body}} when status in [200, 201] ->
@@ -691,8 +512,8 @@ defmodule WandererApp.Esi.ApiClient do
 
         {:ok, %{status: 420, headers: headers} = _error} ->
           # Extract rate limit information from headers
-          reset_seconds = Map.get(headers, "x-esi-error-limit-reset", ["unknown"]) |> List.first()
-          remaining = Map.get(headers, "x-esi-error-limit-remain", ["unknown"]) |> List.first()
+          reset_seconds = Map.get(headers, "x-esi-error-limit-reset", ["0"]) |> List.first()
+          remaining = Map.get(headers, "x-esi-error-limit-remain", ["0"]) |> List.first()
 
           # Emit telemetry for rate limiting
           :telemetry.execute(
@@ -722,6 +543,36 @@ defmodule WandererApp.Esi.ApiClient do
 
           {:error, :error_limited, headers}
 
+        {:ok, %{status: 429, headers: headers} = _error} ->
+          # Extract rate limit information from headers
+          reset_seconds = Map.get(headers, "retry-after", ["0"]) |> List.first()
+
+          # Emit telemetry for rate limiting
+          :telemetry.execute(
+            [:wanderer_app, :esi, :rate_limited],
+            %{
+              count: 1,
+              reset_duration:
+                case Integer.parse(reset_seconds || "0") do
+                  {seconds, _} -> seconds * 1000
+                  _ -> 0
+                end
+            },
+            %{
+              method: "POST_ESI",
+              path: url,
+              reset_seconds: reset_seconds
+            }
+          )
+
+          Logger.warning("ESI_RATE_LIMITED: POST request rate limited",
+            method: "POST_ESI",
+            path: url,
+            reset_seconds: reset_seconds
+          )
+
+          {:error, :error_limited, headers}
+
         {:ok, %{status: status}} ->
           {:error, "Unexpected status: #{status}"}
 
@@ -736,7 +587,7 @@ defmodule WandererApp.Esi.ApiClient do
     end
   end
 
-  defp get_retry(path, api_opts, opts, status \\ :forbidden) do
+  defp do_get_retry(path, api_opts, opts, status \\ :forbidden) do
     refresh_token? = opts |> Keyword.get(:refresh_token?, false)
     retry_count = opts |> Keyword.get(:retry_count, 0)
     character_id = opts |> Keyword.get(:character_id, nil)
@@ -748,7 +599,7 @@ defmodule WandererApp.Esi.ApiClient do
         {:ok, token} ->
           auth_opts = [access_token: token.access_token] |> get_auth_opts()
 
-          get(
+          do_get(
             path,
             api_opts |> Keyword.merge(auth_opts),
             opts |> Keyword.merge(retry_count: retry_count + 1)
@@ -913,44 +764,4 @@ defmodule WandererApp.Esi.ApiClient do
       :character_token_invalid
     )
   end
-
-  defp map_route_info(
-         %{
-           "origin" => origin,
-           "destination" => destination,
-           "systems" => result_systems,
-           "success" => success
-         } = _route_info
-       ),
-       do:
-         map_route_info(%{
-           origin: origin,
-           destination: destination,
-           systems: result_systems,
-           success: success
-         })
-
-  defp map_route_info(
-         %{origin: origin, destination: destination, systems: result_systems, success: success} =
-           _route_info
-       ) do
-    systems =
-      case result_systems do
-        [] ->
-          []
-
-        _ ->
-          result_systems |> Enum.reject(fn system_id -> system_id == origin end)
-      end
-
-    %{
-      has_connection: result_systems != [],
-      systems: systems,
-      origin: origin,
-      destination: destination,
-      success: success
-    }
-  end
-
-  defp map_route_info(_), do: nil
 end
