+++ lib/wanderer_app_web/live/map/map_live.ex
@@ -99,7 +99,7 @@ defmodule WandererAppWeb.MapLive do
        |> push_navigate(to: ~p"/tracking/#{map_slug}")}
 
   @impl true
-  def handle_info(info, %{assigns: %{map_slug: map_slug}} = socket) do
+  def handle_info(info, socket) do
     {:noreply,
      socket
      |> WandererAppWeb.MapEventHandler.handle_event(info)}
